debug

+++ Yunt/Yunt.WebApi/Controllers/UserController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyInjection;
@@ -22,13 +23,21 @@ namespace Yunt.WebApi.Controllers
         public UserController(IUserRepository userRepository)
         {
             _userRepository = userRepository;
-                //ServiceProviderServiceExtensions.GetService<IUserRepository>(Startup.Providers["Auth"]); 
         }
         // GET: api/User
         [HttpGet]
+        [EnableCors("any")]
         public PaginatedList<User> Get(int pageindex,int pagesize)
-        {        
-            return  _userRepository.GetPage(pageindex, pagesize);
+        {
+            try
+            {
+                return _userRepository.GetPage(pageindex, pagesize,null,e=>e.Time);
+            }
+            catch (Exception ex)
+            {
+                Logger.Exception(ex);
+                return new PaginatedList<User>(0, 0, 0, new List<User>());
+            }
         }
 
         // GET: api/User/5
@@ -40,23 +49,58 @@ namespace Yunt.WebApi.Controllers
         
         // POST: api/User
         [HttpPost]
+        [EnableCors("any")]
         public bool Post([FromBody]User value)//dynamic
         {
-             return _userRepository.Insert(value)>0;    
+            try
+            {
+                if (value == null)
+                    return false;
+                value.Time = DateTime.Now.TimeSpan();
+                return _userRepository.Insert(value) > 0;
+            }
+            catch (Exception ex)
+            {
+                Logger.Exception(ex);
+                return false;
+            } 
         }
         
         // PUT: api/User/5
         [HttpPut("{id}")]
+        [EnableCors("any")]
         public bool Put([FromBody]User value)
         {
-            return _userRepository.UpdateEntity(value) > 0;
+            try
+            {
+                if (value == null)
+                    return false;
+                value.Time = DateTime.Now.TimeSpan();
+                return _userRepository.UpdateEntity(value) > 0;
+            }
+            catch (Exception ex)
+            {
+                Logger.Exception(ex);
+                return false;
+            }
         }
         
         // DELETE: api/ApiWithActions/5
         [HttpDelete("{id}")]
         public bool Delete(string id)
         {
-            return _userRepository.DeleteEntity(e => e.UserId.Equals(id)) > 0;
+            try
+            {
+                if (id.IsNullOrWhiteSpace())
+                    return false;
+                return _userRepository.DeleteEntity(e => e.UserId.Equals(id)) > 0;
+            }
+            catch (Exception ex)
+            {
+                Logger.Exception(ex);
+                return false;
+            }
+
         }
     }
 }
